move logic into `utils::playstation`
