+++ relay-server/src/utils/playstation.rs
@@ -1,3 +1,6 @@
+use crate::services::outcome::{DiscardAttachmentType, DiscardItemType};
+use crate::services::processor::ProcessingError;
+use bytes::Bytes;
 use relay_event_schema::protocol::{
     AppContext, ClientSdkInfo, Context, Contexts, DeviceContext, LenientString, OsContext,
     RuntimeContext, Tags,
@@ -154,6 +157,47 @@ pub fn infer_content_type(filename: &str) -> ContentType {
         Some("xml") => ContentType::Xml,
         _ => ContentType::OctetStream,
     }
+}
+
+#[derive(Debug, thiserror::Error)]
+pub enum CompressionError {
+    #[error("failed to decompress dump")]
+    Decode(#[from] std::io::Error),
+    #[error("decompressed dump exceeds size limit")]
+    TooLarge,
+}
+
+impl From<CompressionError> for ProcessingError {
+    fn from(err: CompressionError) -> Self {
+        match err {
+            CompressionError::Decode(err) => {
+                ProcessingError::InvalidPlaystationDump(format!("Failed to extract data: {err}"))
+            }
+            CompressionError::TooLarge => ProcessingError::PayloadTooLarge(
+                DiscardItemType::Attachment(DiscardAttachmentType::Prosperodump),
+            ),
+        }
+    }
+}
+
+pub fn uncompress(bytes: Bytes, limit: usize) -> Result<Bytes, CompressionError> {
+    use std::io::{Cursor, Read};
+
+    const LZ4_MAGIC_HEADER: &[u8] = b"\x04\x22\x4D\x18";
+    if !bytes.starts_with(LZ4_MAGIC_HEADER) {
+        return Ok(bytes);
+    }
+
+    let decoder = lz4_flex::frame::FrameDecoder::new(Cursor::new(bytes));
+    let mut decoder = decoder.take(limit.saturating_add(1) as u64);
+
+    let mut buffer = Vec::new();
+    decoder.read_to_end(&mut buffer)?;
+
+    if buffer.len() > limit {
+        return Err(CompressionError::TooLarge);
+    }
+    Ok(buffer.into())
 }
 
 #[cfg(test)]
